refactor(base): use 128 bit parameters directly instead of downcasting

+++ test/integration/regen/RegenIntegration.t.sol
@@ -193,7 +193,7 @@ contract RegenIntegrationTest is Test {
     function setUp() public virtual {
         rewardTokenDecimals = uint8(bound(vm.randomUint(), 6, 18));
         stakeTokenDecimals = uint8(bound(vm.randomUint(), 6, 18));
-        uint256 rewardDuration = bound(vm.randomUint(), MIN_REWARD_DURATION, MAX_REWARD_DURATION);
+        uint256 rewardDuration = bound(vm.randomUint(), uint128(MIN_REWARD_DURATION), MAX_REWARD_DURATION);
 
         vm.startPrank(ADMIN);
 
@@ -215,7 +215,7 @@ contract RegenIntegrationTest is Test {
             calculator,
             MAX_BUMP_TIP,
             ADMIN,
-            rewardDuration,
+            uint128(rewardDuration),
             MAX_CLAIM_FEE,
             0,
             stakerWhitelist,
@@ -237,7 +237,7 @@ contract RegenIntegrationTest is Test {
     ) public {
         tipAmount = bound(tipAmount, 0, MAX_BUMP_TIP);
         feeAmount = bound(feeAmount, 0, MAX_CLAIM_FEE);
-        minimumStakeAmount = bound(minimumStakeAmount, 0, getStakeAmount(1000));
+        minimumStakeAmount = bound(uint128(minimumStakeAmount), 0, getStakeAmount(1000));
 
         vm.startPrank(ADMIN);
         RegenStaker localRegenStaker = new RegenStaker(
@@ -246,9 +246,9 @@ contract RegenIntegrationTest is Test {
             calculator,
             tipAmount,
             ADMIN,
-            MIN_REWARD_DURATION,
+            uint128(MIN_REWARD_DURATION),
             feeAmount,
-            minimumStakeAmount,
+            uint128(minimumStakeAmount),
             IWhitelist(address(0)),
             IWhitelist(address(0)),
             allocationMechanismWhitelist
@@ -282,7 +282,7 @@ contract RegenIntegrationTest is Test {
     ) public {
         tipAmount = bound(tipAmount, 0, MAX_BUMP_TIP);
         feeAmount = bound(feeAmount, 0, MAX_CLAIM_FEE);
-        minimumStakeAmount = bound(minimumStakeAmount, 0, getStakeAmount(1000));
+        minimumStakeAmount = bound(uint128(minimumStakeAmount), 0, getStakeAmount(1000));
 
         vm.startPrank(ADMIN);
         Whitelist providedStakerWhitelist = new Whitelist();
@@ -297,9 +297,9 @@ contract RegenIntegrationTest is Test {
             calculator,
             tipAmount,
             ADMIN,
-            MIN_REWARD_DURATION,
+            uint128(MIN_REWARD_DURATION),
             feeAmount,
-            minimumStakeAmount,
+            uint128(minimumStakeAmount),
             providedStakerWhitelist,
             providedContributorWhitelist,
             allocationMechanismWhitelist
@@ -352,7 +352,7 @@ contract RegenIntegrationTest is Test {
         newMinimum = bound(newMinimum, 0, getStakeAmount(10000));
 
         vm.prank(ADMIN);
-        regenStaker.setMinimumStakeAmount(newMinimum);
+        regenStaker.setMinimumStakeAmount(uint128(newMinimum));
 
         assertEq(regenStaker.minimumStakeAmount(), newMinimum);
     }
@@ -363,7 +363,7 @@ contract RegenIntegrationTest is Test {
 
         vm.startPrank(nonAdmin);
         vm.expectRevert(abi.encodeWithSelector(Staker.Staker__Unauthorized.selector, bytes32("not admin"), nonAdmin));
-        regenStaker.setMinimumStakeAmount(newMinimum);
+        regenStaker.setMinimumStakeAmount(uint128(newMinimum));
         vm.stopPrank();
     }
 
@@ -372,7 +372,7 @@ contract RegenIntegrationTest is Test {
         stakeAmount = bound(stakeAmount, 1, minimumAmount - 1);
 
         vm.prank(ADMIN);
-        regenStaker.setMinimumStakeAmount(minimumAmount);
+        regenStaker.setMinimumStakeAmount(uint128(minimumAmount));
 
         address user = makeAddr("user");
         whitelistUser(user, true, false, true);
@@ -398,7 +398,7 @@ contract RegenIntegrationTest is Test {
         vm.assume(stakeAmount >= minimumAmount);
 
         vm.prank(ADMIN);
-        regenStaker.setMinimumStakeAmount(minimumAmount);
+        regenStaker.setMinimumStakeAmount(uint128(minimumAmount));
 
         address user = makeAddr("user");
         whitelistUser(user, true, false, true);
@@ -433,7 +433,7 @@ contract RegenIntegrationTest is Test {
         vm.assume(remainingAfterWithdraw + additionalStake < minimumAmount);
 
         vm.prank(ADMIN);
-        regenStaker.setMinimumStakeAmount(minimumAmount);
+        regenStaker.setMinimumStakeAmount(uint128(minimumAmount));
 
         address user = makeAddr("user");
         whitelistUser(user, true, false, true);
@@ -1374,7 +1374,7 @@ contract RegenIntegrationTest is Test {
     }
 
     function testFuzz_SetRewardDuration(uint256 newDuration) public {
-        newDuration = bound(newDuration, MIN_REWARD_DURATION, MAX_REWARD_DURATION);
+        newDuration = bound(newDuration, uint128(MIN_REWARD_DURATION), MAX_REWARD_DURATION);
 
         // current duration
         uint256 currentDuration = regenStaker.rewardDuration();
@@ -1383,18 +1383,18 @@ contract RegenIntegrationTest is Test {
         vm.prank(ADMIN);
         vm.expectEmit(true, true, true, true);
         emit RewardDurationSet(newDuration);
-        regenStaker.setRewardDuration(newDuration);
+        regenStaker.setRewardDuration(uint128(newDuration));
 
         assertEq(regenStaker.rewardDuration(), newDuration);
     }
 
     function testFuzz_RevertIf_NonAdminCannotSetRewardDuration(address nonAdmin, uint256 newDuration) public {
         vm.assume(nonAdmin != ADMIN);
-        newDuration = bound(newDuration, MIN_REWARD_DURATION, MAX_REWARD_DURATION);
+        newDuration = bound(newDuration, uint128(MIN_REWARD_DURATION), MAX_REWARD_DURATION);
 
         vm.startPrank(nonAdmin);
         vm.expectRevert(abi.encodeWithSelector(Staker.Staker__Unauthorized.selector, bytes32("not admin"), nonAdmin));
-        regenStaker.setRewardDuration(newDuration);
+        regenStaker.setRewardDuration(uint128(newDuration));
         vm.stopPrank();
     }
 
@@ -1408,7 +1408,7 @@ contract RegenIntegrationTest is Test {
     function testFuzz_RevertIf_SetRewardDurationToZero() public {
         vm.startPrank(ADMIN);
         vm.expectRevert(abi.encodeWithSelector(RegenStakerBase.InvalidRewardDuration.selector, 0));
-        regenStaker.setRewardDuration(0);
+        regenStaker.setRewardDuration(uint128(0));
         vm.stopPrank();
     }
 
@@ -1420,7 +1420,7 @@ contract RegenIntegrationTest is Test {
     }
 
     function testFuzz_RevertIf_SetRewardDurationDuringActiveReward(uint256 newDuration) public {
-        newDuration = bound(newDuration, MIN_REWARD_DURATION, MAX_REWARD_DURATION);
+        newDuration = bound(newDuration, uint128(MIN_REWARD_DURATION), MAX_REWARD_DURATION);
 
         address staker = makeAddr("staker");
         whitelistUser(staker, true, false, true);
@@ -1442,12 +1442,12 @@ contract RegenIntegrationTest is Test {
 
         vm.startPrank(ADMIN);
         vm.expectRevert(abi.encodeWithSelector(RegenStakerBase.CannotChangeRewardDurationDuringActiveReward.selector));
-        regenStaker.setRewardDuration(newDuration);
+        regenStaker.setRewardDuration(uint128(newDuration));
         vm.stopPrank();
     }
 
     function testFuzz_Constructor_WithCustomRewardDuration(uint256 customDuration) public {
-        customDuration = bound(customDuration, MIN_REWARD_DURATION, MAX_REWARD_DURATION);
+        customDuration = bound(uint128(customDuration), uint128(MIN_REWARD_DURATION), MAX_REWARD_DURATION);
 
         vm.startPrank(ADMIN);
         RegenStaker localRegenStaker = new RegenStaker(
@@ -1456,7 +1456,7 @@ contract RegenIntegrationTest is Test {
             calculator,
             MAX_BUMP_TIP,
             ADMIN,
-            customDuration,
+            uint128(customDuration),
             MAX_CLAIM_FEE,
             0,
             IWhitelist(address(0)),
@@ -1495,10 +1495,10 @@ contract RegenIntegrationTest is Test {
         stakeAmountBase = bound(stakeAmountBase, 1, 100_000);
         customDurationDays = bound(customDurationDays, 30, 365);
         uint256 customDuration = customDurationDays * 1 days;
-        rewardAmountBase = bound(rewardAmountBase, customDuration, 100_000_000);
+        rewardAmountBase = bound(rewardAmountBase, uint128(customDuration), 100_000_000);
 
         vm.prank(ADMIN);
-        regenStaker.setRewardDuration(customDuration);
+        regenStaker.setRewardDuration(uint128(customDuration));
 
         address staker = makeAddr("staker");
         whitelistUser(staker, true, false, true);
@@ -1534,11 +1534,11 @@ contract RegenIntegrationTest is Test {
         stakeAmountBase = bound(stakeAmountBase, 1, 10_000);
         customDurationDays = bound(customDurationDays, 30, 100);
         uint256 customDuration = customDurationDays * 1 days;
-        rewardAmountBase = bound(rewardAmountBase, customDuration, 100_000_000);
+        rewardAmountBase = bound(rewardAmountBase, uint128(customDuration), 100_000_000);
         secondStakerJoinPercent = bound(secondStakerJoinPercent, 10, 90);
 
         vm.prank(ADMIN);
-        regenStaker.setRewardDuration(customDuration);
+        regenStaker.setRewardDuration(uint128(customDuration));
 
         address stakerA = makeAddr("stakerA");
         address stakerB = makeAddr("stakerB");
@@ -1600,11 +1600,11 @@ contract RegenIntegrationTest is Test {
         stakeAmountBase = bound(stakeAmountBase, 1, 10_000);
         customDurationDays = bound(customDurationDays, 30, 100);
         uint256 customDuration = customDurationDays * 1 days;
-        rewardAmountBase = bound(rewardAmountBase, customDuration, 100_000_000);
+        rewardAmountBase = bound(rewardAmountBase, uint128(customDuration), 100_000_000);
         firstClaimTimePercent = bound(firstClaimTimePercent, 10, 90);
 
         vm.prank(ADMIN);
-        regenStaker.setRewardDuration(customDuration);
+        regenStaker.setRewardDuration(uint128(customDuration));
 
         address staker = makeAddr("staker");
         whitelistUser(staker, true, false, true);
@@ -1663,7 +1663,7 @@ contract RegenIntegrationTest is Test {
         rewardAmountBase = bound(rewardAmountBase, max(firstDuration, secondDuration), 100_000_000);
 
         vm.prank(ADMIN);
-        regenStaker.setRewardDuration(firstDuration);
+        regenStaker.setRewardDuration(uint128(firstDuration));
 
         address staker = makeAddr("staker");
         whitelistUser(staker, true, false, true);
@@ -1684,7 +1684,7 @@ contract RegenIntegrationTest is Test {
         vm.warp(block.timestamp + firstDuration + 1);
 
         vm.prank(ADMIN);
-        regenStaker.setRewardDuration(secondDuration);
+        regenStaker.setRewardDuration(uint128(secondDuration));
 
         vm.prank(ADMIN);
         regenStaker.notifyRewardAmount(rewardAmount);
@@ -1728,7 +1728,7 @@ contract RegenIntegrationTest is Test {
             calculator,
             MAX_BUMP_TIP,
             ADMIN,
-            MIN_REWARD_DURATION,
+            uint128(MIN_REWARD_DURATION),
             MAX_CLAIM_FEE,
             0,
             stakerWhitelist,
@@ -1736,7 +1736,7 @@ contract RegenIntegrationTest is Test {
             allocationMechanismWhitelist
         );
         compoundRegenStaker.setRewardNotifier(ADMIN, true);
-        compoundRegenStaker.setMinimumStakeAmount(minimumAmount);
+        compoundRegenStaker.setMinimumStakeAmount(uint128(minimumAmount));
         vm.stopPrank();
 
         address user = makeAddr("user");
@@ -1744,7 +1744,7 @@ contract RegenIntegrationTest is Test {
 
         // Temporarily set minimum to 0 to allow initial stake
         vm.prank(ADMIN);
-        compoundRegenStaker.setMinimumStakeAmount(0);
+        compoundRegenStaker.setMinimumStakeAmount(uint128(0));
 
         sameToken.mint(user, stakeAmount);
         sameToken.mint(address(compoundRegenStaker), rewardAmount);
@@ -1761,7 +1761,7 @@ contract RegenIntegrationTest is Test {
 
         // Reset minimum amount
         vm.prank(ADMIN);
-        compoundRegenStaker.setMinimumStakeAmount(minimumAmount);
+        compoundRegenStaker.setMinimumStakeAmount(uint128(minimumAmount));
 
         uint256 unclaimedBefore = compoundRegenStaker.unclaimedReward(depositId);
         uint256 expectedNewBalance = stakeAmount + unclaimedBefore;
@@ -1789,7 +1789,7 @@ contract RegenIntegrationTest is Test {
         uint256 timeElapsedPercent
     ) public {
         stakeAmountBase = bound(stakeAmountBase, 1, 100_000);
-        rewardAmountBase = bound(rewardAmountBase, MIN_REWARD_DURATION, MAX_REWARD_DURATION + 1_000_000_000);
+        rewardAmountBase = bound(rewardAmountBase, uint128(MIN_REWARD_DURATION), MAX_REWARD_DURATION + 1_000_000_000);
         timeElapsedPercent = bound(timeElapsedPercent, 1, 100);
 
         MockERC20Staking sameToken = new MockERC20Staking(18);
@@ -1801,7 +1801,7 @@ contract RegenIntegrationTest is Test {
             calculator,
             MAX_BUMP_TIP,
             ADMIN,
-            MIN_REWARD_DURATION,
+            uint128(MIN_REWARD_DURATION),
             MAX_CLAIM_FEE,
             0,
             stakerWhitelist,
@@ -1854,7 +1854,7 @@ contract RegenIntegrationTest is Test {
         uint256 feeAmountBase
     ) public {
         stakeAmountBase = bound(stakeAmountBase, 1, 10_000);
-        rewardAmountBase = bound(rewardAmountBase, MIN_REWARD_DURATION, MAX_REWARD_DURATION + 1_000_000_000);
+        rewardAmountBase = bound(rewardAmountBase, uint128(MIN_REWARD_DURATION), MAX_REWARD_DURATION + 1_000_000_000);
         feeAmountBase = bound(feeAmountBase, 0, rewardAmountBase / 10);
 
         MockERC20Staking sameToken = new MockERC20Staking(18);
@@ -1866,7 +1866,7 @@ contract RegenIntegrationTest is Test {
             calculator,
             MAX_BUMP_TIP,
             ADMIN,
-            MIN_REWARD_DURATION,
+            uint128(MIN_REWARD_DURATION),
             MAX_CLAIM_FEE,
             0,
             stakerWhitelist,
@@ -1936,7 +1936,7 @@ contract RegenIntegrationTest is Test {
         uint256 compoundTimes
     ) public {
         stakeAmountBase = bound(stakeAmountBase, 10, 10_000);
-        rewardAmountBase = bound(rewardAmountBase, MIN_REWARD_DURATION, MAX_REWARD_DURATION + 1_000_000_000);
+        rewardAmountBase = bound(rewardAmountBase, uint128(MIN_REWARD_DURATION), MAX_REWARD_DURATION + 1_000_000_000);
         compoundTimes = bound(compoundTimes, 2, 5);
 
         MockERC20Staking sameToken = new MockERC20Staking(18);
@@ -1948,7 +1948,7 @@ contract RegenIntegrationTest is Test {
             calculator,
             MAX_BUMP_TIP,
             ADMIN,
-            MIN_REWARD_DURATION,
+            uint128(MIN_REWARD_DURATION),
             MAX_CLAIM_FEE,
             0,
             stakerWhitelist,
@@ -2008,7 +2008,7 @@ contract RegenIntegrationTest is Test {
         currentTestCtx.user2StakeBase = bound(user2StakeBase, 1, 10_000);
         currentTestCtx.rewardAmountBase = bound(
             rewardAmountBase,
-            MIN_REWARD_DURATION,
+            uint128(MIN_REWARD_DURATION),
             MAX_REWARD_DURATION + 1_000_000_000
         );
         currentTestCtx.user2JoinTimePercent = bound(user2JoinTimePercent, 10, 90);
@@ -2022,7 +2022,7 @@ contract RegenIntegrationTest is Test {
             calculator,
             MAX_BUMP_TIP,
             ADMIN,
-            MIN_REWARD_DURATION,
+            uint128(MIN_REWARD_DURATION),
             MAX_CLAIM_FEE,
             0,
             stakerWhitelist,
@@ -2143,7 +2143,7 @@ contract RegenIntegrationTest is Test {
         uint256 compoundTimePercent
     ) public {
         stakeAmountBase = bound(stakeAmountBase, 1, 10_000);
-        rewardAmountBase = bound(rewardAmountBase, MIN_REWARD_DURATION, MAX_REWARD_DURATION + 1_000_000_000);
+        rewardAmountBase = bound(rewardAmountBase, uint128(MIN_REWARD_DURATION), MAX_REWARD_DURATION + 1_000_000_000);
         compoundTimePercent = bound(compoundTimePercent, 10, 90);
 
         MockERC20Staking sameToken = new MockERC20Staking(18);
@@ -2155,7 +2155,7 @@ contract RegenIntegrationTest is Test {
             calculator,
             MAX_BUMP_TIP,
             ADMIN,
-            MIN_REWARD_DURATION,
+            uint128(MIN_REWARD_DURATION),
             MAX_CLAIM_FEE,
             0,
             stakerWhitelist,
@@ -2212,7 +2212,7 @@ contract RegenIntegrationTest is Test {
         uint8 decimals
     ) public {
         stakeAmountBase = bound(stakeAmountBase, 1, 10_000);
-        rewardAmountBase = bound(rewardAmountBase, MIN_REWARD_DURATION, MAX_REWARD_DURATION + 1_000_000_000);
+        rewardAmountBase = bound(rewardAmountBase, uint128(MIN_REWARD_DURATION), MAX_REWARD_DURATION + 1_000_000_000);
         decimals = uint8(bound(decimals, 6, 18));
 
         MockERC20Staking sameToken = new MockERC20Staking(decimals);
@@ -2224,7 +2224,7 @@ contract RegenIntegrationTest is Test {
             calculator,
             MAX_BUMP_TIP,
             ADMIN,
-            MIN_REWARD_DURATION,
+            uint128(MIN_REWARD_DURATION),
             MAX_CLAIM_FEE,
             0,
             stakerWhitelist,
